Move findNavigateTargetLeaf to core/workspace
Re-homes the navigate-target-leaf resolver from the task-agent adapter
into src/core/workspace so framework code (TerminalPanelView) can share
it without depending on adapter code. Pure rename plus an import path
update in TaskDetailView. No behaviour change.

Relates to #481.
