+++ src/core/workspace/findNavigateTargetLeaf.test.ts
