+++ src/core/workspace/findNavigateTargetLeaf.ts
